[infra] always use valid storage account names
Resource groups are permitted to use dashes, underscores, uppercase letters,
and probably other characters not permitted in storage account names. This
PR cahanges `bootstrap.sh` to:

1. Ignore invalid characters in the resouce group.

2. Ensure (via randomness) that the generated name is unique.

3. Do not try to create a new storage account if `backend-config.tfvars` exists.

I lightly tested this. Here is an example of how it sanitizes a resource group name:

```
RESOURCE_GROUP=bu__ild-batch-worker-i32mage
possibly_invalid_storage_account_name="$(cat /dev/urandom | LC_ALL=C tr -dc 0-9 | head -c 4)${RESOURCE_GROUP}"
STORAGE_ACCOUNT_NAME=$(LC_ALL=C tr -dc a-z0-9 <<< "${possibly_invalid_storage_account_name}" | head -c 24)
echo $STORAGE_ACCOUNT_NAME
7241buildbatchworkeri32m
```

+++ infra/azure/bootstrap.sh
@@ -20,28 +20,35 @@ setup_terraform_backend() {
     RESOURCE_GROUP=$1
     STORAGE_CONTAINER_NAME="tfstate"
 
-    # NOTE: This name is assumed to be globally unique
-    STORAGE_ACCOUNT_NAME="${RESOURCE_GROUP}hailterraform"
+    if [ -e backend-config.tfvars ]
+    then
+        echo "Found backend-config.tfvars. I assume storage account and container for terraform are already created."
+    else
+        echo "No backend-config.tfvars found. I will create storage account and container for terraform."
+        # NOTE: This name is assumed to be globally unique
+        # NOTE: must be 3-24 lowercase letters and numbers only
+        possibly_invalid_storage_account_name="$(cat /dev/urandom | LC_ALL=C tr -dc '0-9' | head -c 4)${RESOURCE_GROUP}"
+        STORAGE_ACCOUNT_NAME=$(LC_ALL=C tr -dc 'a-z0-9' <<< "${possibly_invalid_storage_account_name}" | head -c 24)
 
-    # This storage account/container will be created if it does not already exist
-    az storage account create -n $STORAGE_ACCOUNT_NAME -g $RESOURCE_GROUP
-    STORAGE_ACCOUNT_KEY=$(az storage account keys list \
-        -g $RESOURCE_GROUP \
-        --account-name $STORAGE_ACCOUNT_NAME \
-        | jq -rj '.[0].value'
-    )
+        # This storage account/container will be created if it does not already exist
+        az storage account create -n $STORAGE_ACCOUNT_NAME -g $RESOURCE_GROUP
+        STORAGE_ACCOUNT_KEY=$(az storage account keys list \
+            -g $RESOURCE_GROUP \
+            --account-name $STORAGE_ACCOUNT_NAME \
+            | jq -rj '.[0].value'
+        )
+        az storage container create -n $STORAGE_CONTAINER_NAME \
+            --account-name $STORAGE_ACCOUNT_NAME \
+            --account-key $STORAGE_ACCOUNT_KEY
 
-    az storage container create -n $STORAGE_CONTAINER_NAME \
-        --account-name $STORAGE_ACCOUNT_NAME \
-        --account-key $STORAGE_ACCOUNT_KEY
-
-    cat >backend-config.tfvars <<EOF
+        cat >backend-config.tfvars <<EOF
 storage_account_name = "$STORAGE_ACCOUNT_NAME"
 container_name       = "$STORAGE_CONTAINER_NAME"
 access_key           = "$STORAGE_ACCOUNT_KEY"
 key                  = "haildev.tfstate"
 EOF
 
+    fi
     terraform init -backend-config=backend-config.tfvars
 }
